requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.7.0  # Used for OpenRouter API compatibility
chromadb==0.4.22
sentence-transformers==2.2.2  # For free local embeddings

# Document Processing
pypdf==3.17.4
pdfplumber==0.10.3
python-docx==1.1.0

# API and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# UI
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.26.3
tiktoken==0.5.2

# Monitoring and Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
